+++ pyteal/ast/router.py
@@ -864,7 +864,7 @@ class Router:
 
         return sdk_abi.Contract(self.name, self.methods, self.descr)
 
-    def build_program(
+    def _build_program(
         self,
         *,
         version: int = DEFAULT_TEAL_VERSION,
@@ -915,7 +915,7 @@ class Router:
         bare app calls in the router, and also generates a Contract object to allow client read and call
         the methods easily.
 
-        This method combines :any:`Router.build_program` and :any:`compileTeal`.
+        This method combines `Router._build_program` and :any:`compileTeal`.
 
         Note that if no methods or bare app call actions have been registered to either the approval
         or clear state programs, then that program will reject all transactions.
@@ -929,7 +929,7 @@ class Router:
         """
         starting_slot_id = ScratchSlot.nextSlotId
         try:
-            ap, csp, contract = self.build_program(version=version, optimize=optimize)
+            ap, csp, contract = self._build_program(version=version, optimize=optimize)
             ap_compiled = compileTeal(
                 ap,
                 Mode.Application,